refactored some backend and changed schema
