+++ package.json
@@ -13,6 +13,7 @@
     "@prisma/adapter-pg": "^7.2.0",
     "@prisma/client": "^7.2.0",
     "@types/pg": "^8.16.0",
+    "dotenv": "^17.4.2",
     "motion": "^12.42.2",
     "next": "^16.1.0",
     "pg": "^8.16.3",
