+++ services/astronautService.ts
@@ -1,9 +1,22 @@
-import { getComputedNetWorth, updatePlayerNetWorth } from "./currencyService"
+import { getComputedNetWorth, updateNetWorth } from "./currencyService"
 import { getAstronautView } from "@/views/astronaut"
 import { db } from "../lib/db"
+import { Player } from "@/views/player"
 
 export async function purchaseAstronaut(username: string, astronautName: string) {
     const now = Date.now()
+    const player = await db.user.findUnique({
+        where: { username },
+        include: {
+            astronauts: {
+                include: {
+                    astronautData: true
+                }
+            }
+        }
+    })
+
+    if (!player) throw new Error("Player can not be found.")
 
     const astronautPrice = await db.astronauts.findUniqueOrThrow({
         where: {name : astronautName},
@@ -12,7 +25,8 @@ export async function purchaseAstronaut(username: string, astronautName: string)
         }
     })
 
-    if (astronautPrice.price > await getComputedNetWorth(username, now)) throw new Error("Player cannot afford astronaut.")
+    if (!astronautPrice) throw new Error ("Astronaut cannot be found.")
+    if (astronautPrice.price > await getComputedNetWorth(player, now)) throw new Error("Player cannot afford astronaut.")
 
     let astronaut = await db.ownedAstronauts.create({
         data: {
@@ -26,8 +40,6 @@ export async function purchaseAstronaut(username: string, astronautName: string)
         }
     })
 
-    if (!astronaut) throw new Error()
-
     //toggle idle generation
     if (astronaut.astronautData.isScientist) {
         await db.ownedAstronauts.update({
@@ -38,16 +50,28 @@ export async function purchaseAstronaut(username: string, astronautName: string)
         })
     }
 
-    updatePlayerNetWorth(username, astronaut.astronautData.price * -1, now)
+    updateNetWorth(player, astronaut.astronautData.price * -1, now)
 
     return getAstronautView(astronaut)
 }
 
 export async function sellAstronaut(username: string, astronautId: string) {
     const now = Date.now()
+     const player = await db.user.findUnique({
+        where: { username },
+        include: {
+            astronauts: {
+                include: {
+                    astronautData: true
+                }
+            }
+        }
+    })
+
+    if (!player) throw new Error("Player can not be found.")
     
     //Update player dollar count before deletion to add any dollars the deleted astronaut might have generated
-    await updatePlayerNetWorth(username, 0, now)
+    await updateNetWorth(player, 0, now)
 
     const astronaut = await db.ownedAstronauts.delete({
         where: {id: astronautId},
@@ -58,7 +82,7 @@ export async function sellAstronaut(username: string, astronautId: string) {
 
     if (!astronaut) throw new Error("Astronaut to remove could not be found.")
 
-    await updatePlayerNetWorth(username, astronaut.astronautData.price, now)
+    await updateNetWorth(player, astronaut.astronautData.price, now)
 
     return getAstronautView(astronaut)
 }
